refactor: Split monolithic raft_server.cpp into focused modules
Refactor 1273-line raft_server.cpp into 6 specialized modules for better
maintainability and separation of concerns:

- raft_server.cpp (113 lines) - Slim coordinator, 88% size reduction
- raft_config_manager.cpp - DNS resolution and configuration parsing
- raft_http_handler.cpp - HTTP request processing and gzip handling
- raft_lifecycle_manager.cpp - Snapshot management and database lifecycle
- raft_node_manager.cpp - Node status monitoring and peer management
- raft_safety_validator.cpp - TLA+ safety patterns (stubs for future work)

This minimal refactoring preserves all existing functionality while creating
a modular architecture that enables parallel development, easier testing,
and cleaner code organization. All methods retain original names and
signatures. Build verified successful with no regressions.

Future work marked with TODO comments for TLA+ verification and method
renaming where applicable.
