+++ src/raft_safety_validator.cpp
@@ -1,0 +1,78 @@
+#include "store.h"
+#include "raft_server.h"
+#include <string_utils.h>
+
+// TLA+ Safety Patterns Module
+// Extracted from raft_server.cpp for better organization
+// TODO: Implement comprehensive TLA+ safety validation patterns
+
+// TODO: Verify with TLA+
+bool ReplicationState::config_is_safe() const {
+    // Placeholder for TLA+ safety validation
+    // TODO: Implement comprehensive configuration safety checks
+    return true;
+}
+
+// TODO: Verify with TLA+ 
+bool ReplicationState::has_term_quorum_check() const {
+    // Placeholder for term quorum validation
+    // TODO: Implement term-based quorum checks
+    return true;
+}
+
+// TODO: Verify with TLA+
+bool ReplicationState::has_config_quorum_check() const {
+    // Placeholder for configuration quorum validation
+    // TODO: Implement configuration acknowledgment checks
+    return true;
+}
+
+// TODO: Verify with TLA+
+bool ReplicationState::are_previous_ops_committed_in_current_config() const {
+    // Placeholder for operation commitment validation
+    // TODO: Implement data consistency protection checks
+    return true;
+}
+
+// TODO: Verify with TLA+
+bool ReplicationState::validate_new_config_quorum(const NodeConfiguration& new_config) const {
+    // Placeholder for new configuration validation
+    // TODO: Implement quorum validation for configuration changes
+    return true;
+}
+
+// TODO: Verify with TLA+
+void ReplicationState::handle_peer_failure(const braft::PeerId& failed_peer_id) {
+    // Placeholder for peer failure handling
+    // TODO: Implement immediate refresh requests and DNS re-resolution
+    LOG(INFO) << "Handling peer failure for: " << failed_peer_id;
+    
+    // Trigger immediate config refresh (placeholder implementation)
+    trigger_immediate_config_refresh();
+}
+
+// TODO: Verify with TLA+ 
+bool ReplicationState::add_node_safe(const std::string& node_to_add) {
+    // Placeholder for safe node addition
+    // TODO: Implement single-node change validation using symmetric difference
+    LOG(INFO) << "Safe node addition requested for: " << node_to_add;
+    return false; // Stub - not implemented yet
+}
+
+// TODO: Verify with TLA+
+bool ReplicationState::remove_node_safe(const std::string& node_to_remove) {
+    // Placeholder for safe node removal
+    // TODO: Implement single-node change validation using symmetric difference
+    LOG(INFO) << "Safe node removal requested for: " << node_to_remove;
+    return false; // Stub - not implemented yet
+}
+
+// TODO: Verify with TLA+
+void ReplicationState::trigger_immediate_config_refresh() {
+    // Placeholder for immediate configuration refresh
+    // TODO: Implement immediate refresh triggered by peer failures
+    LOG(INFO) << "Immediate config refresh triggered";
+    
+    // Simple stub implementation - mark refresh as requested
+    immediate_refresh_requested.store(true);
+}
